# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API written in Golang that handles requests from the frontend.",
        "title": "UpAgain backend Golang API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/accounts/{id_account}/pro-analytics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns statistics including inventory, material impact (CO2 and weights), and financial transactions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pro"
                ],
                "summary": "Get professional premium analytics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID of the Professional",
                        "name": "id_account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProAnalyticsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/accounts/{id_account}/pro-analytics/alerts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of materials selected by the pro to receive alerts for.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pro"
                ],
                "summary": "Get professional alert materials",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID of the Professional",
                        "name": "id_account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Saves the list of materials to receive alerts for.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pro"
                ],
                "summary": "Update professional alert materials",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id_account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of materials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProAlertMaterialsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.MaterialInventoryStats": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "integer"
                },
                "available": {
                    "type": "integer"
                },
                "material": {
                    "type": "string"
                },
                "recycled": {
                    "type": "integer"
                }
            }
        },
        "models.MaterialUsageStats": {
            "type": "object",
            "properties": {
                "material": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "models.ProAnalyticsResponse": {
            "type": "object",
            "properties": {
                "finance": {
                    "type": "object",
                    "properties": {
                        "paid_purchases": {
                            "type": "integer"
                        },
                        "total_purchases": {
                            "type": "integer"
                        },
                        "total_spent": {
                            "type": "number"
                        }
                    }
                },
                "impact": {
                    "type": "object",
                    "properties": {
                        "material_usage": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MaterialUsageStats"
                            }
                        },
                        "total_co2": {
                            "type": "number"
                        }
                    }
                },
                "inventory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MaterialInventoryStats"
                    }
                }
            }
        },
...
        "models.UpdateProAlertMaterialsRequest": {
            "type": "object",
            "properties": {
                "materials": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}